test: add tests for different key types, where possible
Expands the tests to test using `secp256k1` and `RSA` keys as well
as `Ed25519` where possible.

Kubo doesn't support generating or importing `secp256k1` keys so
those tests are skipped, as are `RSA` keys for pubsub resolution since
the DHT needs to be enabled to resolve the public key component.